Failed attempt at calling apps synchronously.
The reason I wanted to do this is that I could get rid of the heuristic
5 second pause before re-showing the UI (presumably in the background
because the app should be running by then).

This code hides the UI, launches the app, blocking until the app is
done, then re-shows the UI.

The problem with that is that the whole process hangs, meaning we can't
process any input, which I think will lead to bad behavior when we
return.  Buffers for LIRC, joystick devices, and inotify will all fill
up during that time and I have a strong hunch that will result in ugly
breakage.

So the solution would be to close all of those file descriptors and
re-open them when we return, which I partly implemented.  But then we
have to un-register with glib and re-register and it just seems
obnoxiously over-complicated for something that is probably not even an
issue in the real world.  So I'm abandoning the idea.