Correctly set icon to blank for weird files.
Before, there was a bunch of code in the body of the loop, and the
'icon' variable got reused every iteration, so even though we didn't set
it to anything for irregular file types, the code generally didn't crash
because the variable had some value left over in it.  now, that code is
in a separate function whose namespace starts fresh for every individual
filename, preventing such an oversight from being repeated.