Fullscreen no longer depends on the window manager
The one remaining shortcoming is that we only interrogate the WM upon
startup.  If the WM changes while htpicker is running (including if it
starts an instant after htpicker does, like if they're launched in
parallel and htpicker starts faster), then we will assume the wrong
thing.  The next step is to find a way to monitor for WM changes.

Mostly fixes #15