More concise AND complete.
This fixes the issue where the mouse cursor is visible if it's on the
very edge of the window.  I thought that was some inherent limitation I
was stuck with, but it turns out that the outside 1px (or something) of
the window is indeed part of the document, but not considered part of
the body... or something like that.  In any case, this is a better way
to do it, all around.